Avoid Multiple to_radians Calls
We get the same, or close enough for it to not impact the final result with rounding, when we convert values to radians first, then subtract them. This means that we can avoid an extra to_radians call later. The compiler MAY have also recognized this optimization, but now it is more explicit