Sort Points Prior to Running
It would seem that much of the slowdown when running the program comes from the fact that the branch predictor is wrong more often than not. This may be because the distribution of Points is basically random meaning it ends up with a 50/50 chance of being right. Ordering the points by some filed (in this case longitude), makes the branch predictor much more consistent and results in a ~40% uplift